requirements.txt
anthropic>=0.25.0
chromadb>=0.4.0
sentence-transformers>=2.7.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
junit-xml>=1.9
rich>=13.7.0
typer>=0.12.0
pydantic>=2.7.0
